backend/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
dj-database-url==2.0.0
gunicorn==22.0.0
psycopg2-binary==2.9.9
python-dotenv==1.1.1
sqlparse==0.5.3
